+++ lib/ftpd/session.rb
@@ -147,6 +147,20 @@ module Ftpd
       end
     end
 
+    def cmd_appe(argument)
+      close_data_server_socket_when_done do
+        ensure_logged_in
+        ensure_file_system_supports :append
+        path = argument
+        syntax_error unless path
+        path = File.expand_path(path, @name_prefix)
+        ensure_accessible path
+        contents = receive_file
+        @file_system.append path, contents
+        reply "226 Transfer complete"
+      end
+    end
+
     def cmd_retr(argument)
       close_data_server_socket_when_done do
         ensure_logged_in
@@ -454,7 +468,6 @@ module Ftpd
     end
 
     unimplemented :abor
-    unimplemented :appe
     unimplemented :rein
     unimplemented :rest
     unimplemented :site